+++ src/components/Project/ProjectItem.tsx
@@ -1,7 +1,7 @@
 import MergeRequestList from 'components/MergeRequest/MergeRequestList';
 import React, { Component } from 'react';
 import { Project } from 'types/FormattedTypes';
-import DropDown from 'components/DropDown';
+import Accordion from 'components/Accordion';
 
 interface Props {
   project: Project;
@@ -12,7 +12,7 @@ class ProjectItem extends Component<Props> {
     const { project } = this.props;
 
     return (
-      <DropDown
+      <Accordion
         title={project.name_with_namespace}
         content={(<MergeRequestList mergeRequests={project.mergeRequests} />)}
       />
